text rendering fixes
